multi, singleinputdatefields, drawer, drawermanager, monthpicker stories converted to sb7
